// types/mailchimp-analytics.ts
export interface MailchimpCampaign {
  id: string
  web_id: number
  type: string
  create_time: string
  archive_url: string
  long_archive_url: string
  status: string
  emails_sent: number
  send_time: string
  content_type: string
  needs_block_refresh: boolean
  resendable: boolean
  recipients: {
    list_id: string
    list_is_active: boolean
    list_name: string
    segment_text: string
    recipient_count: number
  }
  settings: {
    subject_line: string
    preview_text: string
    title: string
    from_name: string
    reply_to: string
    use_conversation: boolean
    to_name: string
    folder_id: string
    authenticate: boolean
    auto_footer: boolean
    inline_css: boolean
    auto_tweet: boolean
    fb_comments: boolean
    timewarp: boolean
    template_id: number
    drag_and_drop: boolean
  }
  tracking: {
    opens: boolean
    html_clicks: boolean
    text_clicks: boolean
    goal_tracking: boolean
    ecomm360: boolean
    google_analytics: string
    clicktale: string
  }
  report_summary: {
    opens: number
    unique_opens: number
    open_rate: number
    clicks: number
    subscriber_clicks: number
    click_rate: number
    emails_sent: number
    unsubscribed: number
    bounces: number
    forwards: number
    forward_opens: number
    ecommerce: {
      total_orders: number
      total_spent: number
      total_revenue: number
    }
  }
}
// ...
export interface MailchimpCampaignsResponse {
  data: MailchimpCampaign[]
  totalCampaigns: number
  listId?: string
  _source: string
}